refactor(test): split widget_tests.rs into modular test files
