+++ tests/widget/divider.rs
@@ -1,0 +1,87 @@
+//! Divider widget tests
+
+use revue::layout::Rect;
+use revue::render::Buffer;
+use revue::widget::traits::RenderContext;
+use revue::widget::{divider, vdivider, View};
+
+#[test]
+fn test_divider_render_horizontal() {
+    let mut buffer = Buffer::new(20, 1);
+    let area = Rect::new(0, 0, 20, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = divider();
+    d.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).map(|c| c.symbol), Some('─'));
+}
+
+#[test]
+fn test_divider_render_vertical() {
+    let mut buffer = Buffer::new(1, 10);
+    let area = Rect::new(0, 0, 1, 10);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = vdivider();
+    d.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).map(|c| c.symbol), Some('│'));
+}
+
+#[test]
+fn test_divider_with_label() {
+    let mut buffer = Buffer::new(30, 1);
+    let area = Rect::new(0, 0, 30, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = divider().label("Section");
+    d.render(&mut ctx);
+
+    let text: String = (0..30)
+        .filter_map(|x| buffer.get(x, 0).map(|c| c.symbol))
+        .collect();
+    assert!(text.contains("Section"));
+}
+
+#[test]
+fn test_divider_render_uses_helpers() {
+    let mut buffer = Buffer::new(30, 1);
+    let area = Rect::new(0, 0, 30, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = divider().label("Test");
+    d.render(&mut ctx);
+
+    let text: String = (0..30)
+        .filter_map(|x| buffer.get(x, 0).map(|c| c.symbol))
+        .collect();
+    assert!(text.contains("Test"));
+    assert!(text.contains("─"));
+}
+
+#[test]
+fn test_divider_label_clipping() {
+    let mut buffer = Buffer::new(10, 1);
+    let area = Rect::new(0, 0, 10, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = divider().label("VeryLongLabelThatWontFit");
+    d.render(&mut ctx);
+}
+
+#[test]
+fn test_divider_vertical_uses_vline() {
+    let mut buffer = Buffer::new(1, 5);
+    let area = Rect::new(0, 0, 1, 5);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let d = vdivider();
+    d.render(&mut ctx);
+
+    for y in 0..5 {
+        assert_eq!(buffer.get(0, y).map(|c| c.symbol), Some('│'));
+    }
+}
+
+// =============================================================================
