+++ tests/widget/gauge.rs
@@ -1,0 +1,68 @@
+//! Gauge widget tests
+
+use revue::layout::Rect;
+use revue::render::Buffer;
+use revue::widget::traits::RenderContext;
+use revue::widget::{battery, gauge, percentage, Gauge, GaugeStyle, View};
+
+#[test]
+fn test_gauge_set_get_value() {
+    let mut g = Gauge::new();
+    g.set_value(0.8);
+    assert_eq!(g.get_value(), 0.8);
+}
+
+#[test]
+fn test_gauge_render_all_styles() {
+    let styles = [
+        GaugeStyle::Bar,
+        GaugeStyle::Battery,
+        GaugeStyle::Thermometer,
+        GaugeStyle::Arc,
+        GaugeStyle::Circle,
+        GaugeStyle::Vertical,
+        GaugeStyle::Segments,
+        GaugeStyle::Dots,
+    ];
+
+    for style in styles {
+        let mut buffer = Buffer::new(30, 5);
+        let area = Rect::new(0, 0, 30, 5);
+        let mut ctx = RenderContext::new(&mut buffer, area);
+
+        let g = Gauge::new().style(style).percent(50.0);
+        g.render(&mut ctx);
+    }
+}
+
+#[test]
+fn test_gauge_with_title() {
+    let mut buffer = Buffer::new(30, 3);
+    let area = Rect::new(0, 0, 30, 3);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let g = Gauge::new().title("CPU Usage").percent(75.0);
+    g.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).unwrap().symbol, 'C');
+}
+
+#[test]
+fn test_gauge_helper() {
+    let g = gauge().percent(50.0);
+    assert_eq!(g.get_value(), 0.5);
+}
+
+#[test]
+fn test_percentage_helper() {
+    let g = percentage(75.0);
+    assert_eq!(g.get_value(), 0.75);
+}
+
+#[test]
+fn test_battery_helper() {
+    let g = battery(80.0);
+    assert_eq!(g.get_value(), 0.8);
+}
+
+// =============================================================================
