+++ tests/widget/mod.rs
@@ -1,0 +1,18 @@
+//! Widget integration tests - split into modules by widget type
+
+mod accordion;
+mod alert;
+mod avatar;
+mod badge;
+mod button;
+mod checkbox;
+mod divider;
+mod gauge;
+mod input;
+mod progress;
+mod radio;
+mod spinner;
+mod switch;
+mod tabs;
+mod text;
+mod tooltip;
