+++ tests/widget/text.rs
@@ -1,0 +1,142 @@
+//! Text widget tests
+
+use revue::layout::Rect;
+use revue::render::Buffer;
+use revue::style::Color;
+use revue::widget::traits::RenderContext;
+use revue::widget::{Alignment, StyledView, Text, View};
+
+#[test]
+fn test_text_new() {
+    let text = Text::new("Hello");
+    assert_eq!(text.content(), "Hello");
+}
+
+#[test]
+fn test_text_render() {
+    let mut buffer = Buffer::new(20, 5);
+    let area = Rect::new(0, 0, 20, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let text = Text::new("Hello");
+    text.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).unwrap().symbol, 'H');
+    assert_eq!(buffer.get(1, 0).unwrap().symbol, 'e');
+    assert_eq!(buffer.get(4, 0).unwrap().symbol, 'o');
+}
+
+#[test]
+fn test_text_render_centered() {
+    let mut buffer = Buffer::new(20, 5);
+    let area = Rect::new(0, 0, 20, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    let text = Text::new("Hi").align(Alignment::Center);
+    text.render(&mut ctx);
+
+    assert_eq!(buffer.get(9, 0).unwrap().symbol, 'H');
+    assert_eq!(buffer.get(10, 0).unwrap().symbol, 'i');
+}
+
+#[test]
+fn test_text_css_id() {
+    let text = Text::new("Title").element_id("page-title");
+    assert_eq!(View::id(&text), Some("page-title"));
+
+    let meta = text.meta();
+    assert_eq!(meta.id, Some("page-title".to_string()));
+}
+
+#[test]
+fn test_text_css_classes() {
+    let text = Text::new("Warning").class("alert").class("bold");
+
+    assert!(text.has_class("alert"));
+    assert!(text.has_class("bold"));
+    assert!(!text.has_class("hidden"));
+
+    let meta = text.meta();
+    assert!(meta.classes.contains("alert"));
+    assert!(meta.classes.contains("bold"));
+}
+
+#[test]
+fn test_text_styled_view() {
+    let mut text = Text::new("Test");
+
+    text.set_id("test-text");
+    assert_eq!(View::id(&text), Some("test-text"));
+
+    text.add_class("highlight");
+    assert!(text.has_class("highlight"));
+
+    text.toggle_class("highlight");
+    assert!(!text.has_class("highlight"));
+
+    text.toggle_class("active");
+    assert!(text.has_class("active"));
+
+    text.remove_class("active");
+    assert!(!text.has_class("active"));
+}
+
+#[test]
+fn test_text_css_colors_from_context() {
+    use revue::style::{Style, VisualStyle};
+
+    let text = Text::new("CSS Text");
+    let mut buffer = Buffer::new(20, 3);
+    let area = Rect::new(0, 0, 20, 1);
+
+    let mut style = Style::default();
+    style.visual = VisualStyle {
+        color: Color::MAGENTA,
+        ..VisualStyle::default()
+    };
+
+    let mut ctx = RenderContext::with_style(&mut buffer, area, &style);
+    text.render(&mut ctx);
+}
+
+#[test]
+fn test_text_justify_alignment() {
+    let text = Text::new("Hello World").align(Alignment::Justify);
+    let mut buffer = Buffer::new(20, 1);
+    let area = Rect::new(0, 0, 20, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    text.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).unwrap().symbol, 'H');
+    assert_eq!(buffer.get(4, 0).unwrap().symbol, 'o');
+    assert_eq!(buffer.get(19, 0).unwrap().symbol, 'd');
+    assert_eq!(buffer.get(15, 0).unwrap().symbol, 'W');
+}
+
+#[test]
+fn test_text_justify_single_word() {
+    let text = Text::new("Hello").align(Alignment::Justify);
+    let mut buffer = Buffer::new(20, 1);
+    let area = Rect::new(0, 0, 20, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    text.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).unwrap().symbol, 'H');
+    assert_eq!(buffer.get(4, 0).unwrap().symbol, 'o');
+}
+
+#[test]
+fn test_text_justify_multiple_words() {
+    let text = Text::new("A B C").align(Alignment::Justify);
+    let mut buffer = Buffer::new(11, 1);
+    let area = Rect::new(0, 0, 11, 1);
+    let mut ctx = RenderContext::new(&mut buffer, area);
+
+    text.render(&mut ctx);
+
+    assert_eq!(buffer.get(0, 0).unwrap().symbol, 'A');
+    assert_eq!(buffer.get(5, 0).unwrap().symbol, 'B');
+    assert_eq!(buffer.get(10, 0).unwrap().symbol, 'C');
+}
